% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clear_labels.R
\name{clear_labels}
\alias{clear_labels}
\title{Removes Hmisc labels from a dataframe}
\usage{
clear_labels(x)
}
\arguments{
\item{x}{A data.frame.}
}
\description{
Some functions do not work when variables in a data.frame are having the 
"Labelled" class appended to them. This removes labels from a dataframe.
}
\examples{
library(Hmisc)
getHdata(plasma)
#Look-up labels:
label(plasma)

#Remove labels:
clean_plasma <- clear_labels(plasma)
label(clean_plasma)
}
\keyword{labels}
\keyword{remove}

