% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_labels.R
\name{define_labels}
\alias{define_labels}
\title{Define Hmisc labels from a list}
\usage{
define_labels(x, labels)
}
\arguments{
\item{x}{A data.frame.}

\item{labels}{A list containing label information, e.g. var = "Variable name"}
}
\description{
Apply labels to several variables simultaneously from a definition found in a
list object.
}
\examples{
library(Hmisc)
getHdata(plasma)
#Look-up labels:
label(plasma)

#Remove labels:
clean_plasma <- clear_labels(plasma)
label(clean_plasma)

#Define variable labels from a list:
var_names <- list(age = "Age in years", smokstat = "Smoking status")
named_plasma <- define_labels(clean_plasma, var_names)
label(named_plasma)
}

