% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowchart.R
\name{flowchart}
\alias{flowchart}
\title{Create a flowchart}
\usage{
flowchart(description, n = NULL, events = NULL)
}
\arguments{
\item{description}{Typically a character vector with text for each of the boxes in the flowchart.}

\item{n}{Numeric vector of same length as 'description', describing number of subjects in each group.}

\item{events}{Numeric vector of same length, describing number of events.}
}
\description{
This function will draw a flowchart, useful for many purposes, e.g. visualizing
the population selection for a specific study, and many others. If n and events
are provided, event rates will automatically populate the boxes in the flowchart.
}
\examples{
flow_data <- data.frame(descr = c("Population", "Subset1", "Further subset"),
                        n = c(1000, 800, 600), events = c(80, 70 , 30))
with(flow_data, flowchart(descr, n, events))
}
\author{
Daniel Lindholm
}
\keyword{flowchart}

