% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hrci.R
\name{hrci}
\alias{hrci}
\title{Combine hazard ratio and 95 percent CI to one string}
\usage{
hrci(f_data, hr = hr, lci = lci, uci = uci)
}
\arguments{
\item{f_data}{a dataframe containing information for the forestplot}

\item{hr}{hazard ratio; hr by default}

\item{lci}{lower confidence interval; lci by default}

\item{uci}{upper confidence interval; uci by default}
}
\description{
Useful for combining HR and 95CI information from a dataframe used for 
drawing a forestplot.
}
\examples{
Create some test data
hr <- seq(0.80, 1.1, by = 0.05)
lci <- hr - 0.1
uci <- hr + 0.1
left_text<- paste0("Left text no ", 1:7)
forest_data <- data.frame(left_text,lci,hr,uci)

forest_data$right_text <- hrci(forest_data)
forestplot(forest_data, hr, lci, uci, left_text, right_text) 
}
\keyword{forestplot}

