% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\name{info}
\alias{info}
\title{Describe selection performed with dplyr functions}
\usage{
info(data, descr_df = character(), descr_text = character(), event = NULL)
}
\arguments{
\item{data}{A data.frame.}

\item{descr_df}{A character string specifying where to store the descriptive information.}

\item{descr_text}{A character string describing the population.}

\item{event}{A character string indicating which column of the data.frame contains information on
whether an event occured. If this argument is specified, the number of events 
is saved to descriptive information.}
}
\description{
When selecting an analysis population using the dplyr
functions (e.g. filter), it could be useful to describe
each step of the way. This function allows text annotation and
provides number of observations for each step.
}
\details{
Important! The descr_df should be non-existent in the beginning of the selection
procedures!
}
\examples{
library(dplyr)
aq <- airquality

aq \%>\% info("descr", "Full dataset") \%>\% 
filter(Temp < 60) \%>\% info("descr", "Temperature below 60")

print(descr)
}
\keyword{describe}
\keyword{selection,}

