% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metafix.R
\name{metafix}
\alias{metafix}
\title{Remove attributes from a data.frame}
\usage{
metafix(df)
}
\arguments{
\item{df}{a dataframe to fix}
}
\description{
When loading e.g. SAS datasets using the haven package also loads irritating attributes like
SAS labels. This makes dplyr functions unhappy. The present function removes these attributes.
}

