% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_plot.R
\name{missing_plot}
\alias{missing_plot}
\title{Visualize missing values in a data.frame}
\usage{
missing_plot(data, group_variable, fill_color = "steelblue")
}
\arguments{
\item{data}{a data.frame}

\item{group_variable}{a character string naming the column of the data.frame that 
contains information on grouping (e.g. treatment group, etc)}

\item{fill_color}{the plot will show a gradient from white to this color.
By default, steelblue will be used.}
}
\description{
This function looks at missing values in a data.frame in relation to a specific
group and creates a ggplot2 object. Quite helpful to identify missing not-at-random.
}
\examples{

missing_plot(airquality, "Month")

missing_plot(airquality, "Month", "Orange")
}
\keyword{data}
\keyword{missing}
\keyword{missing_plot,}

