% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_table.R
\name{missing_table}
\alias{missing_table}
\title{Create table showing missing values}
\usage{
missing_table(data, group_variable = NULL, round_values = TRUE,
  show_total = FALSE, show_non_missing = TRUE)
}
\arguments{
\item{data}{A data.frame.}

\item{group_variable}{A character string. If missing values should be assessed 
in relation to a specific group, this argument specifies the name of the column
which contains grouping information.}

\item{round_values}{Should the numbers be rounded?}

\item{show_total}{When looking at a table of missing values according to group,
should missingness in the whole population also be showed?}

\item{show_non_missing}{Should information about missingness be shown for all
variables in the data sets (even those with no missing values)?}
}
\description{
A tool to assess missingness in a dataset. This can be expanded to look into
specific groups of the data. Returns a data.frame.
}
\examples{
missing_table(airquality)

missing_table(airquality, show_non_missing=FALSE)

missing_table(airquality, "Month", show_total = TRUE)
}
\keyword{missing_table}

