% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_table.R
\name{save_table}
\alias{save_table}
\title{Wrapper to write dataframe to disk as a csv and use chunk name}
\usage{
save_table(df, table_name = opts_current$get("label"), directory = "tables",
  row_names = FALSE)
}
\arguments{
\item{df}{a dataframe}

\item{table_name}{the name of the file to write. By default, it will be same as
the label of the chunk from which it was called}

\item{directory}{the directory in which the file will be written}

\item{row_names}{include rownames in the .csv-file? FALSE by default.}
}
\description{
This function is a wrapper for write_csv. It will write a dataframe to disk.
If no table_name is provided and the function is called from within a knitr
code chunk, the table_name variable will take the same value as the chunk label.
By default, the directory is a 'tables' subdirectory to the working directory,
which suits my directory structure nicely. Could of course be customized.
}
\author{
Daniel Lindholm
}

