% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selection_flowcart.R
\name{selection_flowchart}
\alias{selection_flowchart}
\title{Draw a flowchart of study population selection}
\usage{
selection_flowchart(description = data.frame())
}
\arguments{
\item{description}{A description table created with the describe function.}
}
\description{
When using the dplyr package and the info function from the current package,
the resulting population selection table can be used to draw a flowchart with
the present function
}
\examples{
library(dplyr)
library(datasets)
# A simple example
aq_subset <- airquality \%>\% 
               info("descr", "Full dataset") \%>\% 
               filter(Temp < 60) \%>\% 
               info("descr", "Temperature below 60")
selection_flowchart(descr)

# An example with Event information as well
library(survival)
heart_subset <- heart \%>\% 
                   info("descr2", "Full study population", "event") \%>\% 
                   filter(transplant == 1) \%>\% 
                   info("descr2", "Transplanted patients", "event")
selection_flowchart(descr2)
}
\keyword{describe}
\keyword{selection,}

