% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variable_table.R
\name{variable_table}
\alias{variable_table}
\title{Create variable table}
\usage{
variable_table(df)
}
\arguments{
\item{df}{a data.frame}
}
\description{
Create a table in the form of a data.frame, which contains information about
variable name, variable description (as defined using the label function from
the Hmisc package), and class of variable. This could then of course be used together with
the latex function (in Hmisc) or kable (in knitr) to put the table into reports etc.
}
\examples{
library(Hmisc)
getHdata(plasma)
variable_table(plasma)
}
\author{
Daniel Lindholm
}

